+++ tests/unit/test_experiment.py
@@ -2,13 +2,14 @@
 """Unit tests for the experiment.py."""
 
 from pathlib import Path
+from unittest.mock import Mock, patch
 
 import pytest
 import tomli
 import tomlkit
 
 from deode.config_parser import ParsedConfig
-from deode.experiment import ExpFromFiles
+from deode.experiment import ExpFromFiles, case_setup
 
 
 @pytest.fixture(name="tmpdir", scope="module")
@@ -79,24 +80,66 @@ def fixture_config(default_config):
     return config
 
 
-def test_exp_from_nonexisting_file(
-    config: ParsedConfig, nonexisting_file: Path, exp_dependencies: dict
-):
-    """Test function for creating an experiment from a non-existing file."""
-    ExpFromFiles(config, exp_dependencies, [nonexisting_file])
+class TestExpFromFiles:
+    """Unit tests for the ExpFromFiles class."""
+
+    def test_exp_from_nonexisting_file(
+        self, config: ParsedConfig, nonexisting_file: Path, exp_dependencies: dict
+    ):
+        """Test function for creating an experiment from a non-existing file."""
+        ExpFromFiles(config, exp_dependencies, [nonexisting_file])
+
+    def test_exp_from_nontoml_file(
+        self, config: ParsedConfig, nontoml_file: Path, exp_dependencies: dict
+    ):
+        """Test function for creating an experiment from a non-TOML file."""
+        with pytest.raises((RuntimeError, tomli.TOMLDecodeError)):
+            ExpFromFiles(config, exp_dependencies, [nontoml_file])
+
+    def test_exp_from_toml_file(
+        self,
+        config: ParsedConfig,
+        toml_file: Path,
+        exp_dependencies: dict,
+        output_file: Path,
+    ):
+        """Test function for creating an experiment from a TOML file."""
+        exp = ExpFromFiles(config, exp_dependencies, [toml_file])
+        exp.config.save_as(output_file)
 
 
-def test_exp_from_nontoml_file(
-    config: ParsedConfig, nontoml_file: Path, exp_dependencies: dict
-):
-    """Test function for creating an experiment from a non-TOML file."""
-    with pytest.raises((RuntimeError, tomli.TOMLDecodeError)):
-        ExpFromFiles(config, exp_dependencies, [nontoml_file])
+@patch("deode.experiment.EPSExp")
+class TestCaseSetup:
+    """Unit tests for the case_setup function."""
 
+    @patch("deode.experiment.ExpFromFiles")
+    def test_eps_not_activated(
+        self,
+        exp_from_files: Mock,
+        epsexp_mock: Mock,
+        config: ParsedConfig,
+        output_file: Path,
+    ):
+        """Test that EPS setup is not activated."""
+        exp_from_files.config = config
 
-def test_exp_from_toml_file(
-    config: ParsedConfig, toml_file: Path, exp_dependencies: dict, output_file: Path
-):
-    """Test function for creating an experiment from a TOML file."""
-    exp = ExpFromFiles(config, exp_dependencies, [toml_file])
-    exp.config.save_as(output_file)
+        case_setup(config=config, output_file=output_file, mod_files=[])
+
+        epsexp_mock.assert_not_called()
+
+    @patch("deode.experiment.ExpFromFiles")
+    def test_eps_activated(
+        self,
+        epsexp_mock: Mock,
+        exp_from_files: Mock,
+        config: ParsedConfig,
+        output_file: Path,
+    ):
+        """Test that EPS setup is activated."""
+        # Add EPS configuration to the main configuration to activate EPS
+        config = config.copy(update={"eps": {}})
+        exp_from_files.config = config
+
+        case_setup(config=config, output_file=output_file, mod_files=[])
+
+        epsexp_mock.assert_called_once()